Removed range inputs. Moved 'apply' button.
Range inputs don't really offer a whole lot of value and they sort of
clutter the ui. So I removed them. Also, since the 'apply' button is
only relevant when you've selected dates on the calendar pickers, I
grouped it with the calendars and it will only be displayed when they
are displayed.